Changed start activity to be OpenDaysActivity
also removed main activity
