+++ app/src/main/AndroidManifest.xml
@@ -9,14 +9,13 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".InstituteInfoActivity"></activity>
+        <activity
+            android:name=".InstituteInfoActivity"
+            android:parentActivityName=".OpenDaysActivity" />
         <activity
             android:name=".OpenDayActivity"
             android:parentActivityName=".OpenDaysActivity" />
-        <activity
-            android:name=".OpenDaysActivity"
-            android:parentActivityName=".MainActivity" />
-        <activity android:name=".MainActivity">
+        <activity android:name=".OpenDaysActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
